Break bidirectional dependency
